Dump recordings into temporary directory if test path name does not resolve to file.
In Quarkus (tested version: 2.0.0.Alpha2), the path to the test source
file can resolve to a URI with a scheme of “quarkus”, which does not
designate a real file system path.  In this case, Path#of throws a
FileSystemNotFoundException, which we now catch and use to fall back
to a directory created with Files#createTempDirectory.